+++ decider-bcl2fastq/CHANGELOG.md
@@ -2,6 +2,7 @@
 - [GP-1211](https://jira.oicr.on.ca/browse/GP-1121) - Support for single end sequencing
 - [GP-1900](https://jira.oicr.on.ca/browse/GP-1900) - bcl2fastq option to not provision out undetermined fastqs
 - [GP-1914](https://jira.oicr.on.ca/browse/GP-1914) - Fix filtering of samples
+- [GP-1901](https://jira.oicr.on.ca/browse/GP-1901) - Use workflowType to determine lane splitting mode
 ## 1.2.4 - 2018-12-04
 - [GP-1883](https://jira.oicr.on.ca/browse/GP-1883) - Only retrieve analysis records from "check-wf-accessions"
 ## 1.2.3 - 2018-11-20